pyproject.toml
[tool.poetry]
name = "safari"
version = "1.2.0"
description = "Python utilities for Safari."
authors = ["Yevgnen Koh <wherejoystarts@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytoml = "^0.1.21"
PyYAML = "^5.4.1"
lz4 = "^3.1.3"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
jedi-language-server = "0.21.0"
nose2 = "^0.10.0"
pre-commit = "^2.10.0"
pylint = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
